refactor(mitm-addon): retire legacy model JSON fallback
