+++ crates/runner/mitm-addon/src/usage/__init__.py
@@ -30,9 +30,7 @@ helpers remain private.
 
 from . import webhook
 from .anthropic_messages import (
-    create_anthropic_messages_json_usage_extractor,
     create_anthropic_messages_sse_usage_extractor,
-    extract_anthropic_messages_usage_with_error_from_json,
 )
 from .buffer import (
     DEFAULT_FLUSH_INTERVAL_SECONDS,
@@ -57,13 +55,9 @@ from .model_json import (
     ModelJsonResponseInspection,
     ModelUsageProtocol,
     create_model_json_response_inspector,
-    create_model_json_usage_extractor,
-    extract_model_usage_with_error_from_json,
 )
 from .openai_chat_completions import (
-    create_openai_chat_completions_json_usage_extractor,
     create_openai_chat_completions_sse_usage_extractor,
-    extract_openai_chat_completions_usage_with_error_from_json,
 )
 from .openai_responses import (
     OPENAI_RESPONSES_WEBSOCKET_WORK_LIMIT_ERROR,
@@ -72,10 +66,8 @@ from .openai_responses import (
     OpenAIResponsesServerEventInspection,
     OpenAIResponsesServerFailureEvidence,
     OpenAIResponsesServerLifecycle,
-    create_openai_responses_json_usage_extractor,
     create_openai_responses_sse_usage_extractor,
     extract_openai_responses_usage_from_event,
-    extract_openai_responses_usage_with_error_from_json,
     inspect_openai_responses_client_event_json,
     inspect_openai_responses_event_json,
     inspect_openai_responses_server_event,
@@ -112,23 +104,15 @@ __all__ = [
     "buffer_source_usage_events",
     "buffer_usage_events",
     "configure_usage_buffer",
-    "create_anthropic_messages_json_usage_extractor",
     "create_anthropic_messages_sse_usage_extractor",
     "create_connector_response_parser",
     "create_model_json_response_inspector",
-    "create_model_json_usage_extractor",
-    "create_openai_chat_completions_json_usage_extractor",
     "create_openai_chat_completions_sse_usage_extractor",
-    "create_openai_responses_json_usage_extractor",
     "create_openai_responses_sse_usage_extractor",
     "current_usage_state_id",
     "decrement_in_flight_flows",
     "drain_usage_events_after_executor_shutdown",
-    "extract_anthropic_messages_usage_with_error_from_json",
-    "extract_model_usage_with_error_from_json",
-    "extract_openai_chat_completions_usage_with_error_from_json",
     "extract_openai_responses_usage_from_event",
-    "extract_openai_responses_usage_with_error_from_json",
     "flush_usage_events",
     "has_connector_response_parser",
     "has_positive_model_provider_usage",
