+++ crates/runner/mitm-addon/src/usage/model_json.py
@@ -1,15 +1,8 @@
-"""Typed cross-provider dispatch for model JSON usage extraction.
-
-Each supported protocol owns one paired registration containing its incremental
-extractor factory and bounded complete-body extractor. Cross-provider callers
-must dispatch through this module so the two JSON paths cannot drift.
-"""
+"""Shared cross-provider model JSON response inspection."""
 
 from collections.abc import Callable, Mapping
 from dataclasses import dataclass
-from typing import Literal, NamedTuple, Protocol, assert_never
-
-from mitmproxy import http
+from typing import Literal, NamedTuple, assert_never
 
 from . import anthropic_messages, openai_chat_completions, openai_responses
 from .json_selective import JsonExtractionResult, JsonSelectiveExtractor, ScalarField
@@ -30,29 +23,7 @@ ModelUsageProtocol = Literal[
 ModelJsonUsageResult = tuple[dict | None, str | None]
 
 
-class ModelJsonUsageExtractor(Protocol):
-    """Incremental parser lifecycle shared by model JSON protocols."""
-
-    def feed(self, chunk: bytes) -> None:
-        raise NotImplementedError
-
-    def accepts_more_input(self) -> bool:
-        raise NotImplementedError
-
-    def finish(self) -> ModelJsonUsageResult:
-        raise NotImplementedError
-
-
-_ModelJsonUsageExtractorFactory = Callable[[], ModelJsonUsageExtractor]
-_ModelJsonUsageCompleteBodyExtractor = Callable[
-    [bytes, http.Headers | None],
-    ModelJsonUsageResult,
-]
-
-
 class _ModelJsonUsageRegistration(NamedTuple):
-    create_extractor: _ModelJsonUsageExtractorFactory
-    extract_from_complete_body: _ModelJsonUsageCompleteBodyExtractor
     scalar_fields: Callable[[], Mapping[JsonPath, ScalarField]]
     object_presence_paths: Callable[[], set[JsonPath]]
     value_presence_paths: Callable[[], set[JsonPath]]
@@ -60,30 +31,18 @@ class _ModelJsonUsageRegistration(NamedTuple):
 
 
 _ANTHROPIC_MESSAGES_REGISTRATION = _ModelJsonUsageRegistration(
-    create_extractor=anthropic_messages.create_anthropic_messages_json_usage_extractor,
-    extract_from_complete_body=(
-        anthropic_messages.extract_anthropic_messages_usage_with_error_from_json
-    ),
     scalar_fields=anthropic_messages.model_json_scalar_fields,
     object_presence_paths=set,
     value_presence_paths=set,
     usage_from_result=anthropic_messages.model_json_usage_from_result,
 )
 _OPENAI_CHAT_COMPLETIONS_REGISTRATION = _ModelJsonUsageRegistration(
-    create_extractor=(openai_chat_completions.create_openai_chat_completions_json_usage_extractor),
-    extract_from_complete_body=(
-        openai_chat_completions.extract_openai_chat_completions_usage_with_error_from_json
-    ),
     scalar_fields=openai_chat_completions.model_json_scalar_fields,
     object_presence_paths=openai_chat_completions.model_json_object_presence_paths,
     value_presence_paths=openai_chat_completions.model_json_value_presence_paths,
     usage_from_result=openai_chat_completions.model_json_usage_from_result,
 )
 _OPENAI_RESPONSES_REGISTRATION = _ModelJsonUsageRegistration(
-    create_extractor=openai_responses.create_openai_responses_json_usage_extractor,
-    extract_from_complete_body=(
-        openai_responses.extract_openai_responses_usage_with_error_from_json
-    ),
     scalar_fields=openai_responses.model_json_scalar_fields,
     object_presence_paths=set,
     value_presence_paths=set,
@@ -244,21 +203,3 @@ def _model_json_usage_registration(
     if protocol == "openai_responses":
         return _OPENAI_RESPONSES_REGISTRATION
     return assert_never(protocol)
-
-
-def create_model_json_usage_extractor(
-    protocol: ModelUsageProtocol,
-) -> ModelJsonUsageExtractor:
-    """Create the incremental JSON usage extractor for ``protocol``."""
-
-    return _model_json_usage_registration(protocol).create_extractor()
-
-
-def extract_model_usage_with_error_from_json(
-    protocol: ModelUsageProtocol,
-    body: bytes,
-    headers: http.Headers | None,
-) -> ModelJsonUsageResult:
-    """Extract usage from one complete, optionally encoded model JSON body."""
-
-    return _model_json_usage_registration(protocol).extract_from_complete_body(body, headers)
